x,y: int --> double

+++ IAD2/Points/Center.h
@@ -6,10 +6,12 @@
 class Center : public Point
 {
 public:
-    Center(int x, int y) ;
+    Center(double x, double y) ;
 
     void reposition_center( std::vector<Point> allPoints ) ;
     std::vector<Point> return_belonging_points( std::vector<Point> allPoints ) ;
+
+
 };
